Enhance parallel processing and memory management in chunk execution; refactor full_length_quantification function and update related tests
